+++ libs/profiling/CHANGELOG.rst
@@ -1,3 +1,7 @@
+**0.5.3 - 07/16/26**
+
+- Stop xfailing two tests
+
 **0.5.2 - 07/14/26**
 
 - Update stale references for monorepo libraries